7_django/Django Accounts Article/requirements.txt
appnope==0.1.3
asgiref==3.5.2
asttokens==2.0.8
backcall==0.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
decorator==5.1.1
django-extensions==3.2.0
executing==0.10.0
idna==3.3
ipython==8.4.0
jedi==0.18.1
matplotlib-inline==0.1.6
mypy-extensions==0.4.3
mysqlclient==2.1.1
numpy==1.23.2
pandas==1.4.3
parso==0.8.3
pathspec==0.10.1
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.5.2
prompt-toolkit==3.0.30
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
PyMySQL==1.0.2
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.4
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
stack-data==0.4.0
tomli==2.0.1
traitlets==5.3.0
typing_extensions==4.4.0
tzdata==2022.4
urllib3==1.26.10
wcwidth==0.2.5
